refactor: hooks to fetch block number
